[Update] UpperBound_LowerBound / Abstract Function
