+++ FS/__basic.py
@@ -1,0 +1,32 @@
+import numpy as np
+from numpy.random import rand
+
+
+def init_position(lb, ub, N, dim):
+    X = np.zeros([N, dim], dtype='float')
+    for i in range(N):
+        for d in range(dim):
+            X[i,d] = lb[0,d] + (ub[0,d] - lb[0,d]) * rand()        
+    
+    return X
+
+
+def binary_conversion(X, thres, N, dim):
+    Xbin = np.zeros([N, dim], dtype='int')
+    for i in range(N):
+        for d in range(dim):
+            if X[i,d] > thres:
+                Xbin[i,d] = 1
+            else:
+                Xbin[i,d] = 0
+    
+    return Xbin
+
+
+def boundary(x, lb, ub):
+    if x < lb:
+        x = lb
+    if x > ub:
+        x = ub
+    
+    return x
