+++ FS/ga.py
@@ -1,28 +1,7 @@
 import numpy as np
 from numpy.random import rand
 from FS.functionHO import Fun
-
-
-def init_position(lb, ub, N, dim):
-    X = np.zeros([N, dim], dtype='float')
-    for i in range(N):
-        for d in range(dim):
-            X[i,d] = lb[0,d] + (ub[0,d] - lb[0,d]) * rand()        
-    
-    return X
-
-
-def binary_conversion(X, thres, N, dim):
-    Xbin = np.zeros([N, dim], dtype='int')
-    for i in range(N):
-        for d in range(dim):
-            if X[i,d] > thres:
-                Xbin[i,d] = 1
-            else:
-                Xbin[i,d] = 0
-    
-    return Xbin
-
+from FS.__basic import init_position, binary_conversion, boundary
 
 def roulette_wheel(prob):
     num = len(prob)
@@ -38,8 +17,8 @@ def roulette_wheel(prob):
 
 def jfs(xtrain, ytrain, opts):
     # Parameters
-    ub       = 1
-    lb       = 0
+    ub       = opts['ub']
+    lb       = opts['lb']
     thres    = 0.5    
     CR       = 0.8     # crossover rate
     MR       = 0.01    # mutation rate
