+++ FS/pso.py
@@ -1,15 +1,7 @@
 import numpy as np
 from numpy.random import rand
 from FS.functionHO import Fun
-
-
-def init_position(lb, ub, N, dim):
-    X = np.zeros([N, dim], dtype='float')
-    for i in range(N):
-        for d in range(dim):
-            X[i,d] = lb[0,d] + (ub[0,d] - lb[0,d]) * rand()        
-    
-    return X
+from FS.__basic import init_position, binary_conversion, boundary
 
 
 def init_velocity(lb, ub, N, dim):
@@ -26,33 +18,12 @@ def init_velocity(lb, ub, N, dim):
             V[i,d] = Vmin[0,d] + (Vmax[0,d] - Vmin[0,d]) * rand()
         
     return V, Vmax, Vmin
-
-
-def binary_conversion(X, thres, N, dim):
-    Xbin = np.zeros([N, dim], dtype='int')
-    for i in range(N):
-        for d in range(dim):
-            if X[i,d] > thres:
-                Xbin[i,d] = 1
-            else:
-                Xbin[i,d] = 0
-    
-    return Xbin
-
-
-def boundary(x, lb, ub):
-    if x < lb:
-        x = lb
-    if x > ub:
-        x = ub
-    
-    return x
     
 
 def jfs(xtrain, ytrain, opts):
     # Parameters
-    ub    = 1
-    lb    = 0
+    ub    = opts['ub']
+    lb    = opts['lb']
     thres = 0.5
     w     = 0.9    # inertia weight
     c1    = 2      # acceleration factor
